feat: added support for parenthesis priority in query
done task(20260703-022205)

+++ lib/task.h
@@ -14,7 +14,7 @@ typedef struct {
     char *path;
     char *uuid;
     char *name;
-    uint8_t priority;
+    size_t priority;
     task_status status;
     Ht(const char *, bool) tags;
 } task_t;
